MB-58486: Adding Resumability when collection gets added/removed
So Mongo Connector works in two stages
1. A snapshot is being taken of the included collections individually.
2. Once snapshot is complete for all the collections, a opLog change stream
   starts, and the filtering is done based on the involved collections.
3. The filtered records are pushed into kafka topic, and resumeToken is recorded
   in the kafka offset topic. so that when the connector is being restarted it
   resumes from the point it left.

The issue was when the connector is deleted, and a new collection gets added for
scanning. Debezium connector is skipping the snapshot part as it already have a resume resumeToken
based on the previous run, and directly starts streaming from the change log along with the older collections.
so, the older data of the newly added collection was not getting pushed into the respective kafka topic.

In order to address that this patch make some changes to the how resumeToken are getting recorded
in the offset topic. Initially resumeToken was getting recorded at connector level, this patch aimed
to stored resumeToken at the collection level.

current offset Pattern:
  key = ["A5",{"rs":"Cluster0-shard-0","server_id":"Sideline"}]
  value = {"sec":0,"ord":-1,"transaction_id":null,"resume_token":"826503F048000000052B0229296E04"}

new offset Pattern:
  key = ["A5",{"rs":"Cluster0-shard-0","server_id":"Sideline"}]
  value = {"sec":1694757815,"ord":3,"transaction_id":null,"__collection_sample_airbnb.listingsAndReviews":"826503F2510000000A2B022C0100296E5A10042837DD8C15DD4F04B387021C939D6358463C5F6964003C3130303832343232000004","resume_token":"826503F3B7000000032B022C0100296E5A1004C892DD08E68041529B33DA83C8EA3EFD463C5F6964003C363364636536393030393062366534366265333230303866000004","__collection_sample_mflix.movies":"826503F2B3000000082B0229296E04","__collection_sample_airbnb.friends":"826503F3B7000000032B022C0100296E5A1004C892DD08E68041529B33DA83C8EA3EFD463C5F6964003C363364636536393030393062366534366265333230303866000004"}

Change-Id: I046c34cd828269d7be2f9247ac1d0afa9457204b
Reviewed-on: https://review.couchbase.org/c/kafka-connect-mongo/+/197229
Reviewed-by: Hussain Towaileb <Hussain.Towaileb@Couchbase.com>
Tested-by: Ritik Raj <raj.ritik9835@gmail.com>